requirements.txt
# Day-Night Transition Simulation
# Required Python packages

PyOpenGL==3.1.7
PyOpenGL-accelerate==3.1.7
Pillow>=10.0.0
numpy>=1.24.0
